Removed prefacing slashes from URLs.
The docs (http://www.dataprotocols.org/en/latest/data-packages.html) were updated to show that a URL is intended to be an absolute refence to a file which may or may not be on this domain, whereas a path was intended to be a local, relative reference from the directory in which the datapckage.json file was hosted. Since ours are in the same dir, I just used the file name directly.